requirements.txt
customtkinter
openai
python-dotenv
playwright
google-cloud-speech
sounddevice==0.4.6
numpy==1.26.4
cffi==1.16.0
pyinstaller
langchain
langchain-openai
langchain-community
pypdf
chromadb
unstructured
python-dotenv
tiktoken
watchdog
streamlit
unstructured[xlsx]
langchain-unstructured
selenium
webdriver-manager